requirements.txt
python-dateutil==2.6.0
requests==2.13.0
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.8
django-filter==1.0.1
django-markdown-deux==1.0.5
djangorestframework==3.5.4
gunicorn==19.6.0
Markdown==2.6.8
packaging==16.8
psycopg2==2.6.2
pyparsing==2.2.0
six==1.10.0
whitenoise==3.3.0
CommonMark==0.7.3
